Gradually moving stories to its component
